+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - 2025-07-25
+### Changed
+
+ - Use `frozenlist` to de-serialize lists (sequences, arrays)
+ - Use `frozendict` to de-serialize maps
+ - Use `namedtuple` to de-serialize records
+
 ## [0.1.1] - 2025-07-25
 ### Added